+++ src/backend/worksheet/plots/AbstractPlotPrivate.h
@@ -39,8 +39,8 @@ public:
 	virtual QString name() const;
 	virtual void retransform() {}
 
-	float horizontalPadding; //horiz. offset between the plot area and the area defining the coordinate system, in scene units
-	float verticalPadding; //vert. offset between the plot area and the area defining the coordinate system, in scene units
+	double horizontalPadding; //horiz. offset between the plot area and the area defining the coordinate system, in scene units
+	double verticalPadding; //vert. offset between the plot area and the area defining the coordinate system, in scene units
 	double rightPadding;
 	double bottomPadding;
 	bool symmetricPadding;
